Updating rule around removing version prefix
Multiple languages have rules around how to generate version numbers
which include more than just the `major`.`minor`.`patch` set up of a
version. This could include various suffixes like `dev`, `alpha`,
`beta`, and `rc` or even the ability to tag a version without all three
items in semantic versioning.

To support these various version rules, we need to be more accepting of
alternate tagging structures. This commit adds in the ability to tag
with items such as `v1`, `v1.1`, `v1.alpha`, `v1.1.1.beta` and still
strip the `v` from the beginning of the tag in addition to the original
rule that was put in place which was limited to only allowing `v1.1.1`
type tags the ability to strip the `v`.